requirements.txt
django==3.2.7
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
